bug fix and add cf feature

+++ lmdo/cmds/cf/cloudformation.py
@@ -6,6 +6,8 @@ import datetime
 import time
 import shutil
 
+from botocore.exceptions import ClientError
+
 from lmdo.cmds.aws_base import AWSBase
 from lmdo.cmds.s3.s3 import S3
 from lmdo.oprint import Oprint
@@ -49,7 +51,7 @@ class Cloudformation(AWSBase):
         """Create/Update stack"""
         # Don't run if we don't have templates
         if not self._config.get('CloudFormation'):
-            Oprint.info('No cloudformation found, skip', 'cloudformation')
+            Oprint.info('No cloudformation found, skip', self.NAME)
             return True
 
         self.process()
@@ -57,10 +59,15 @@ class Cloudformation(AWSBase):
     def delete(self):
         """Delete stack"""
         if not self._config.get('CloudFormation'):
-            Oprint.info('No cloudformation found, skip', 'cloudformation')
+            Oprint.info('No cloudformation found, skip', self.NAME)
             return True
 
         for stack in self._config.get('CloudFormation').get('Stacks'):
+            # If only for specified stack
+            specified_stack = self.if_specify_stack()
+            if specified_stack and specified_stack != stack.get('Name'):
+                continue
+
             stack_name = self.get_lmdo_format_name(stack.get('Name'), stack.get('DisablePrefix', False))
             self.delete_stack(stack_name)
 
@@ -68,10 +75,14 @@ class Cloudformation(AWSBase):
         """Wrapper, same action as create"""
         self.create()
 
+    def if_specify_stack(self):
+        """If user specify a stack to process"""
+        return False if not self._args.get('--stack') else self._args.get('--stack')
+
     def prepare(self, templates, bucket=None):
         """Prepare all templates/validate/upload before create and update"""
         if len(templates['children']) > 0 and not bucket:
-            Oprint.err('S3 bucket hasn\'t been provided for nested template', 'cloudformation')
+            Oprint.err('S3 bucket hasn\'t been provided for nested template', self.NAME)
 
         # Validate syntax of the template
         with open(templates['master'], 'r') as outfile:
@@ -98,7 +109,7 @@ class Cloudformation(AWSBase):
         try:
             result = self._client.validate_template(TemplateBody=template_body)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
         return True
 
     def get_stack(self, stack_name, cached=False):
@@ -117,6 +128,11 @@ class Cloudformation(AWSBase):
             s3_bucket = self._config.get('CloudFormation').get('S3Bucket')
             repo_path = self._config.get('CloudFormation').get('TemplateRepoPath')
             for stack in self._config.get('CloudFormation').get('Stacks'):
+                # If only for specified stack
+                specified_stack = self.if_specify_stack()
+                if specified_stack and specified_stack != stack.get('Name'):
+                    continue
+
                 func_params = {}
 
                 params_path = stack.get('ParamsPath')
@@ -135,7 +151,7 @@ class Cloudformation(AWSBase):
                 if stack_info:
                     # You cannot update a stack with status ROLLBACK_COMPLETE during creation
                     if stack_info['Stacks'][0]['StackStatus'] == 'ROLLBACK_COMPLETE':
-                        Oprint.warn('Stack {} exited with bad state ROLLBACK_COMPLETE. Required to be removed first'.format(stack_name), 'cloudformation')
+                        Oprint.warn('Stack {} exited with bad state ROLLBACK_COMPLETE. Required to be removed first'.format(stack_name), self.NAME)
                         self.delete_stack(stack_name, no_policy=True)
                     else:
                         to_update = True
@@ -183,7 +199,7 @@ class Cloudformation(AWSBase):
 
             self.lock_stack(stack_name=stack_name)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
             return False
 
         self.verify_stack(mode='create', stack_id=response.get('StackId'))
@@ -195,8 +211,6 @@ class Cloudformation(AWSBase):
         try:
             capabilities = capabilities or ['CAPABILITY_NAMED_IAM', 'CAPABILITY_IAM']
             self.unlock_stack(stack_name=stack_name)
-
-            from botocore.exceptions import ClientError
 
             if self._args.get('-e') or self._args.get('--event'):
                 response = self._client.update_stack(
@@ -222,7 +236,7 @@ class Cloudformation(AWSBase):
             else:
                 Oprint.err(str(ce.response['Error']['Message']), self.NAME)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
             return False
 
         self.verify_stack(mode='update', stack_id=response.get('StackId'))
@@ -248,7 +262,7 @@ class Cloudformation(AWSBase):
                 response = self._client.delete_stack(StackName=stack_name)
                 waiter.wait(stack_name)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
             return False
 
         self.verify_stack(mode='delete', stack_id=stack_info['Stacks'][0]['StackId'])
@@ -274,15 +288,15 @@ class Cloudformation(AWSBase):
 
         if mode == 'create':
             if status != 'CREATE_COMPLETE':
-                Oprint.err("Create stack failed with status {}".format(status), 'cloudformation')
+                Oprint.err("Create stack failed with status {}".format(status), self.NAME)
 
         if mode == 'update':
             if status != 'UPDATE_COMPLETE':
-                Oprint.err("Update stack failed with status {}".format(status), 'cloudformation')
+                Oprint.err("Update stack failed with status {}".format(status), self.NAME)
 
         if mode == 'delete':
             if status != 'DELETE_COMPLETE':
-                Oprint.warn("Delete stack failed with status {} (most likely caused by your S3 buckets have contents)".format(status), 'cloudformation')
+                Oprint.warn("Delete stack failed with status {} (most likely caused by your S3 buckets have contents)".format(status), self.NAME)
 
     def get_output_value(self, stack_name, key, cached=False):
         """get a specific stack output value"""
@@ -300,7 +314,7 @@ class Cloudformation(AWSBase):
         try:
             response = self._client.list_change_sets(StackName=stack_name, *args, **kwargs)
         except Exception as e:
-            Oprint.warn(e, 'cloudformation')
+            Oprint.warn(e, self.NAME)
 
         return response
 
@@ -309,7 +323,7 @@ class Cloudformation(AWSBase):
         try:
             response = self._client.describe_change_set(ChangeSetName=change_set_name, *args, **kwargs)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
 
         return response
 
@@ -323,7 +337,7 @@ class Cloudformation(AWSBase):
         try:
             response = self._client.delete_change_set(ChangeSetName=change_set_name, *args, **kwargs)
         except Exception as e:
-            Oprint.warn(e, 'cloudformation')
+            Oprint.warn(e, self.NAME)
 
         return True
 
@@ -334,12 +348,12 @@ class Cloudformation(AWSBase):
 
             change_set_name = self.create_change_set_name(stack_name)
 
-            #Oprint.info('Creating change set {} for stack {}'.format(change_set_name, stack_name), 'cloudformation')
+            #Oprint.info('Creating change set {} for stack {}'.format(change_set_name, stack_name), self.NAME)
             response = self._client.create_change_set(StackName=stack_name, ChangeSetName=change_set_name, *args, **kwargs)
 
             waiter.wait(change_set_name=change_set_name, stack_name=stack_name)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
 
         return change_set_name
 
@@ -349,19 +363,19 @@ class Cloudformation(AWSBase):
             self.unlock_stack(stack_name=stack_name)
 
             if self._args.get('-e') or self._args.get('--event'):
-                Oprint.info('Executing change set {} for updating stack {}'.format(change_set_name, stack_name), 'cloudformation')
+                Oprint.info('Executing change set {} for updating stack {}'.format(change_set_name, stack_name), self.NAME)
                 response = self._client.execute_change_set(ChangeSetName=change_set_name, StackName=stack_name, *args, **kwargs)
 
                 self.stack_events_waiter(stack_name=stack_name)
             else:
                 waiter = CloudformationWaiterStackUpdate(self._client)
-                Oprint.info('Executing change set {} for updating stack {}'.format(change_set_name, stack_name), 'cloudformation')
+                Oprint.info('Executing change set {} for updating stack {}'.format(change_set_name, stack_name), self.NAME)
                 response = self._client.execute_change_set(ChangeSetName=change_set_name, StackName=stack_name, *args, **kwargs)
                 waiter.wait(stack_name)
 
             self.lock_stack(stack_name=stack_name)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
 
         self.verify_stack(mode='update', stack_id=stack_name)
 
@@ -374,10 +388,12 @@ class Cloudformation(AWSBase):
             with open(lock_policy, 'r') as outfile:
                 policy = outfile.read()
 
-            Oprint.info('Locking stack {} to prevent accidental changes'.format(stack_name), 'cloudformation')
+            Oprint.info('Locking stack {} to prevent accidental changes'.format(stack_name), self.NAME)
             response = self._client.set_stack_policy(StackName=stack_name, StackPolicyBody=policy)
+        except ClientError as ce:
+            Oprint.err(str(ce.response['Error']['Message']), self.NAME)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
 
         return True
 
@@ -388,10 +404,10 @@ class Cloudformation(AWSBase):
             with open(unlock_policy, 'r') as outfile:
                 policy = outfile.read()
 
-            Oprint.info('Unlocking stack {} for update'.format(stack_name), 'cloudformation')
+            Oprint.info('Unlocking stack {} for update'.format(stack_name), self.NAME)
             response = self._client.set_stack_policy(StackName=stack_name, StackPolicyBody=policy)
         except Exception as e:
-            Oprint.err(e, 'cloudformation')
+            Oprint.err(e, self.NAME)
 
         return True
 
@@ -399,7 +415,7 @@ class Cloudformation(AWSBase):
         try:
             response = self._client.describe_stack_events(StackName=stack_name, *args, **kwargs)
         except Exception as e:
-            Oprint.warn(e, 'cloudformation')
+            Oprint.warn(e, self.NAME)
         return response
 
     def display_stack_event(self, stack_name, *args, **kwargs):
@@ -416,7 +432,7 @@ class Cloudformation(AWSBase):
                 event['ResourceStatus'],
                 event.get('ResourceStatusReason', '')
             ])
-            Oprint.info('{}{}'.format(prefix, event_info), 'cloudformation')
+            Oprint.info('{}{}'.format(prefix, event_info), self.NAME)
 
             # Update timestamp
             if not self.current_event_timestamp:
@@ -428,9 +444,12 @@ class Cloudformation(AWSBase):
         """Event waiter"""
         in_progress = True
         while in_progress:
-            in_progress = self.get_stack_status(stack_id=stack_name, status_niddle=CfStatus.STACK_IN_PROGRESS)
-            self.display_stack_event(stack_name=stack_name)
-            time.sleep(3)
+            try:
+                in_progress = self.get_stack_status(stack_id=stack_name, status_niddle=CfStatus.STACK_IN_PROGRESS)
+                self.display_stack_event(stack_name=stack_name)
+                time.sleep(3)
+            except Exception:
+                in_progress = False
 
         return in_progress
 
@@ -440,7 +459,7 @@ class Cloudformation(AWSBase):
         # If no changes, skip
         if not change_set_info.get('Changes'):
             print(change_set_info)
-            Oprint.warn('There is no changes in change set, abort', 'cloudformation')
+            Oprint.warn('There is no changes in change set, abort', self.NAME)
             return False
 
         self.pretty_change_set_changes(change_set_info.get('Changes'))
@@ -456,28 +475,28 @@ class Cloudformation(AWSBase):
 
     def pretty_change_set_changes(self, changes):
         """Display changes in a pretty format"""
-        Oprint.infog('Proposed changes', 'cloudformation')
+        Oprint.infog('Proposed changes', self.NAME)
         for item in changes:
             change = item['ResourceChange']
-            Oprint.infog('========================', 'cloudformation')
-            Oprint.infog('Type: {}'.format(item.get('Type')), 'cloudformation')
-            Oprint.infog('Action: {}'.format(change.get('Action')), 'cloudformation')
-            Oprint.infog('Replacement: {}'.format(change.get('Replacement')), 'cloudformation')
-            Oprint.infog('Scope: {}'.format(', '.join(change.get('Scope') or [])), 'cloudformation')
-            Oprint.infog('ResourceType: {}'.format(change.get('ResourceType')), 'cloudformation')
-            Oprint.infog('LogicalResourceId: {}'.format(change.get('LogicalResourceId')), 'cloudformation')
-            Oprint.infog('PhysicalResourceId: {}'.format(change.get('PhysicalResourceId')), 'cloudformation')
-            Oprint.infog('Details:', 'cloudformation')
+            Oprint.infog('========================', self.NAME)
+            Oprint.infog('Type: {}'.format(item.get('Type')), self.NAME)
+            Oprint.infog('Action: {}'.format(change.get('Action')), self.NAME)
+            Oprint.infog('Replacement: {}'.format(change.get('Replacement')), self.NAME)
+            Oprint.infog('Scope: {}'.format(', '.join(change.get('Scope') or [])), self.NAME)
+            Oprint.infog('ResourceType: {}'.format(change.get('ResourceType')), self.NAME)
+            Oprint.infog('LogicalResourceId: {}'.format(change.get('LogicalResourceId')), self.NAME)
+            Oprint.infog('PhysicalResourceId: {}'.format(change.get('PhysicalResourceId')), self.NAME)
+            Oprint.infog('Details:', self.NAME)
 
             for detail in change['Details']:
-                Oprint.infog('    ChangeSource: {}'.format(detail.get('ChangeSource')), 'cloudformation')
-                Oprint.infog('    Evaluation: {}'.format(detail.get('Evaluation')), 'cloudformation')
-                Oprint.infog('    CausingEntity: {}'.format(detail.get('CausingEntity')), 'cloudformation')
-                Oprint.infog('    Target:', 'cloudformation')
-                Oprint.infog('        Attribute: {}'.format(detail.get('Target').get('Attribute')), 'cloudformation')
-                Oprint.infog('        Name: {}'.format(detail.get('Target').get('Name')), 'cloudformation')
-                Oprint.infog('        RequiresRecreation: {}'.format(detail.get('Target').get('RequiresRecreation')), 'cloudformation')
-                Oprint.infog('        -------------------------', 'cloudformation')
+                Oprint.infog('    ChangeSource: {}'.format(detail.get('ChangeSource')), self.NAME)
+                Oprint.infog('    Evaluation: {}'.format(detail.get('Evaluation')), self.NAME)
+                Oprint.infog('    CausingEntity: {}'.format(detail.get('CausingEntity')), self.NAME)
+                Oprint.infog('    Target:', self.NAME)
+                Oprint.infog('        Attribute: {}'.format(detail.get('Target').get('Attribute')), self.NAME)
+                Oprint.infog('        Name: {}'.format(detail.get('Target').get('Name')), self.NAME)
+                Oprint.infog('        RequiresRecreation: {}'.format(detail.get('Target').get('RequiresRecreation')), self.NAME)
+                Oprint.infog('        -------------------------', self.NAME)
 
     def stack_update_via_change_set(self, stack_name, *args, **kwargs):
         """Securely update a stack"""
